refactor(cli): extract pr-merge into commands/pr_commands.py
Consolidates two conflicting pr-merge definitions:
- Removed dead Click stub at cli.py:2523 (shadowed at runtime by
  cli.add_command at line 5697 and never reachable).
- Removed live argparse delegation from cli.py:5679-5697.
- Both now live in src/dopemux/commands/pr_commands.py as a single
  argparse-backed pr_merge_group, registered once via
  cli.add_command near the native-hooks registration.

Preserves _register_routing_commands() as the last statement before
main(). CliRunner smoke-test of `pr-merge --help` returns exit 0 with
the specialist's argparse usage output. CLI test suite shows the same
4 pre-existing failures (test_cli.py renamed-command drift); no new
regressions.
